pyproject.toml
[tool.poetry]
name = "dropcoin"
version = "0.1.0"
description = ""
authors = ["FourZd <shogo.ryaskov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.1"
uvicorn = "^0.29.0"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.2.1"
sqlalchemy = "^2.0.29"
httpx = "^0.27.0"
tweepy = "^4.14.0"
alembic = "^1.13.1"
asyncpg = "^0.29.0"
psycopg2 = "^2.9.9"
python-jose = "^3.3.0"
pillow = "^10.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
